requirements.txt
python-dotenv
requests
beautifulsoup4
serpapi
ollama
python-telegram-bot
python-telegram-bot[ext]